C++ enhancements (#78)
* Refactored use of the Head/BaseClass structure so that is C++ friendly
* CPP refactoring across all modules

+++ src/picture_jpeg/lib/jmemansi.c
@@ -72,7 +72,7 @@ void read_backing_store(j_common_ptr cinfo, backing_store_ptr info, void FAR * b
 
 void write_backing_store(j_common_ptr cinfo, backing_store_ptr info, void FAR * buffer_address, long file_offset, long byte_count)
 {
-   struct acSeek seek = { .Offset = file_offset, .Position = SEEK_START };
+   struct acSeek seek = { .Offset = (DOUBLE)file_offset, .Position = SEEK_START };
    if (Action(AC_Seek, info->temp_file, &seek) != ERR_Okay) ERREXIT(cinfo, JERR_TFILE_SEEK);
 
    struct acWrite write = { .Buffer = buffer_address, .Length = byte_count };
